refactor: adjust aztec.js note access related tests doc and variable name updates
