+++ packages/note-access/src/index.js
@@ -1,5 +1,7 @@
-import { METADATA_AZTEC_DATA_LENGTH, AZTEC_JS_METADATA_PREFIX_LENGTH, VIEWING_KEY_LENGTH } from './config/constants';
+import { AZTEC_JS_METADATA_PREFIX_LENGTH, VIEWING_KEY_LENGTH } from './config/constants';
 import metadata from './metadata';
 import generateAccessMetaData from './generateAccessMetaData';
 
-export { AZTEC_JS_METADATA_PREFIX_LENGTH, generateAccessMetaData, metadata, METADATA_AZTEC_DATA_LENGTH, VIEWING_KEY_LENGTH };
+const constants = { AZTEC_JS_METADATA_PREFIX_LENGTH, VIEWING_KEY_LENGTH };
+
+export { constants, generateAccessMetaData, metadata};
